añadidos decimales a la hora de añadir saldo a la cuenta

+++ Django_app/clubby/forms/forms_miguel.py
@@ -34,13 +34,11 @@ class TicketPurchaseForm(forms.Form):
 
 
 class FundsForm(forms.Form):
-    ammount = forms.IntegerField(
-        max_value=500, min_value=10, help_text=_("how much of your currency you want to add, max 500€, min 10€"))
+    ammount = forms.DecimalField(decimal_places=2,max_value=500, min_value=10, help_text=_("how much of your currency you want to add, max 500€, min 10€"))
 
 
 class PremiumForm(forms.Form):
-    accept = forms.BooleanField(
-        help_text=_("if you agree with these terms we welcome you to the clubby team!"))
+    accept = forms.BooleanField(help_text=_("if you agree with these terms we welcome you to the clubby team!"))
 
 
 class SearchForm(forms.Form):
